requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
torch>=2.1.0
torchvision>=0.16.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pyIGRF>=12.0
netCDF4>=1.6.4
filterpy>=1.4.5
scikit-learn>=1.3.0
websockets>=12.0
onnx>=1.14.0
onnxruntime>=1.16.0
requests>=2.31.0
psutil>=5.9.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
h5py>=3.9.0
matplotlib>=3.7.0
tqdm>=4.66.0
numba>=0.58.0
